requirements.txt
Flask==3.0.0
mysql-connector-python==9.1.0
boto3==1.34.0
prometheus-flask-exporter==0.23.0
gunicorn==22.0.0
python-dotenv==1.0.0

# gunicorn==22.0.0                  # was 21.2.0 — fixes CVE-2024-1135 and CVE-2024-6827
# mysql-connector-python==9.1.0     # was 8.3.0  — fixes CVE-2024-21272
